Added peripheral and main layers for LPC4078
- Shifted main from entry_point to l5_application
- Changed main.c to CONTROLLER_NAME_main
- Added uart driver to peripherals
-
