+++ Template/Minimal_Interfaces/l2_peripheral/LPC4078xx/LPC4078xx.cmake
@@ -1,0 +1,9 @@
+set(l2_peripheral_sources "")
+set(l2_peripheral_dirs "")
+
+ess_data(l2_peripheral_sources 
+    peripheral.c
+)
+ess_data(l2_peripheral_dirs
+    .
+)
