+++ Template/Minimal_Interfaces/l5_application/STM32L475xx_main.c
