+++ Template/Minimal_Interfaces/l5_application/l5_application.cmake
@@ -2,7 +2,7 @@ set(l5_application_sources "")
 set(l5_application_dirs "")
 
 ess_data(l5_application_sources 
-    main.c
+    ${CONTROLLER_NAME}_main.c
 )
 ess_data(l5_application_dirs
     .
